Refactored Grid type accordingly to changes of MosaicShape trait
Renamed TiltedGrid to Grid type
Removed shearing of shape because this feature will be available for all mosaic shapes
