+++ src/mosaic_shape/mod.rs
@@ -169,11 +169,11 @@ impl Clone for Box<dyn MosaicShape> {
     }
 }
 
+mod grid;
 mod helpers;
 mod polygonal_star;
 mod regular_polygon;
-mod tilted_grid;
 
+pub use grid::Grid;
 pub use polygonal_star::PolygonalStar;
 pub use regular_polygon::RegularPolygon;
-pub use tilted_grid::TiltedGrid;
